arregle la ubicacion de los componentes
