+++ src/components/Error404.jsx
@@ -24,3 +24,5 @@ export const Error404 = () =>{
         </Container>
     )
 }
+
+export default Error404
