+++ src/components/views/Colaboradores.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import '../App.css';
+import '../../App.css';
 import { Container, Carousel } from 'react-bootstrap';
 
 const Colaboradores = () => {
