--- src/components/views/Contacto.jsx
@@ -1,167 +1,0 @@
-import { Button, Col, Container, FloatingLabel, Form, Row, InputGroup } from "react-bootstrap"
-import { useForm } from "react-hook-form";
-
-export const Contacto = () => {
-    const { register, handleSubmit, formState: { errors }, reset } = useForm();
-
-    const onSubmit = () => {
-        reset();
-    }
-
-    return (
-        <Container fluid className="mt-5">
-            <Row className="mt-5">
-                <Col xs={{ span: 8, offset: 2 }} className="p-3 bg-body-tertiary mt-5">
-                    <div className="p-3 text-center">
-                        <h1 className="fw-bold">CONTACTANOS</h1>
-                        <div className="m-3 text-center">
-                            <img className="my-3" src="https://cristiangramajo.vercel.app/_next/static/media/main.87519bdd.svg" alt="Contacto rolling vet" />
-                        </div>
-                        <p>Ya sea email, llamada, o completa el formulario para saber como podemos ayudarte con tu consulta y planificar a tu gusto</p>
-                        <small><i><b>*Para urgencias recomendamos que llame a la linea habiltada 24hs</b></i></small>
-                        <ul style={{ listStyle: 'none' }}>
-                            <li>
-                                <a href="mailto:cristiangramajo015@gmail.com" className="link-dark link-offset-2 link-offset-3-hover link-underline link-underline-opacity-0 link-underline-opacity-75-hover">rollingvet@gmail.com</a>
-                            </li>
-                            <li>
-                                <a href="tel:+543865332311" className="link-dark link-offset-2 link-offset-3-hover link-underline link-underline-opacity-0 link-underline-opacity-75-hover">+54 (3865) 332311</a>
-                            </li>
-                            <li>
-                                <a href="#" className="link-warning link-offset-2 link-offset-3-hover link-underline link-underline-opacity-0 link-underline-opacity-75-hover">Portal de soporte, chat online.</a>
-                            </li>
-                        </ul>
-                    </div>
-                </Col>
-                <Col xs={{ span: 8, offset: 2 }} className="p-3 mt-5">
-                    <h4 className="text-center fw-bold">Formulario de contacto</h4>
-                    <Form className="p-3 bg-white border rounded" onSubmit={handleSubmit(onSubmit)}>
-                        <Row className="mb-3">
-                            <Col xs={12} md={6} lg={6} className="mt-3" >
-                                <Form.Group controlId="nombre">
-                                    <Form.Label>Nombre</Form.Label>
-                                    <Form.Control type="text" placeholder="Ingrese su nombre" {...register("nombre", {
-                                        required: "El nombre es un dato obligatorio",
-                                        minLength: {
-                                            value: 2,
-                                            message: "La cantidad minima de caracteres es de 2 digitos",
-                                        },
-                                        maxLength: {
-                                            value: 30,
-                                            message: "La cantidad maxima de caracteres es de 30 digitos",
-                                        },
-                                    })} />
-                                    <Form.Text className="text-danger">
-                                        {errors.nombre?.message}
-                                    </Form.Text>
-                                </Form.Group>
-                            </Col>
-
-                            <Col xs={12} md={6} lg={6} className="mt-3" >
-                                <Form.Group controlId="apellido">
-                                    <Form.Label>Apellido</Form.Label>
-                                    <Form.Control type="text" placeholder="Ingrese su apellido" {...register("apellido", {
-                                        required: "El apeliido es un dato obligatorio",
-                                        minLength: {
-                                            value: 2,
-                                            message: "La cantidad minima de caracteres es de 2 digitos",
-                                        },
-                                        maxLength: {
-                                            value: 30,
-                                            message: "La cantidad maxima de caracteres es de 30 digitos",
-                                        },
-                                    })} />
-                                    <Form.Text className="text-danger">
-                                        {errors.apellido?.message}
-                                    </Form.Text>
-                                </Form.Group>
-                            </Col>
-                            <Col xs={12} md={6} lg={6} className="mt-3" >
-                                <Form.Group controlId="email">
-                                    <Form.Label>Email</Form.Label>
-                                    <Form.Control type="email" placeholder="name@email.com"
-                                        {...register("email", {
-                                            required: "El email es un dato obligatorio",
-                                            pattern: {
-                                                value:
-                                                    /^[a-z0-9!#$%&'*+/=?^_`{|}~-]+(?:\.[a-z0-9!#$%&'*+/=? ^_`{|}~-]+)*@(?:[a-z0-9](?:[a-z0-9-]*[a-z0-9])?\.)+[a -z0-9](?:[a-z0-9-]*[a-z0-9])?$/,
-                                                message:
-                                                    "El email debe tener el siguiente formato: mail@dominio.com",
-                                            },
-                                        })} />
-                                    <Form.Text className="text-danger">
-                                        {errors.email?.message}
-                                    </Form.Text>
-                                </Form.Group>
-                            </Col>
-                            <Col xs={12} md={6} lg={6} className="mt-3" >
-                                <Form.Group controlId="nombreMasco">
-                                    <Form.Label>Nombre de mascota</Form.Label>
-                                    <Form.Control type="text" placeholder="Ingrese el nombre de la mascota" {...register("nombreMasco", {
-                                        required: "El nombre de la mascota es un dato obligatorio",
-                                        minLength: {
-                                            value: 2,
-                                            message: "La cantidad minima de caracteres es de 2 digitos",
-                                        },
-                                        maxLength: {
-                                            value: 30,
-                                            message: "La cantidad maxima de caracteres es de 30 digitos",
-                                        },
-                                    })} />
-                                    <Form.Text className="text-danger">
-                                        {errors.nombreMasco?.message}
-                                    </Form.Text>
-                                </Form.Group>
-                            </Col>
-                            <Col xs={12} md={12} lg={12} className="mt-3" >
-                                    <Form.Label>Telefono</Form.Label>
-                                <InputGroup controlId="telefono" className="mb-3">
-                                    <InputGroup.Text id="telefono">+54</InputGroup.Text>
-                                    <Form.Control type="number" placeholder="+543865332311" aria-describedby="telefono"
-                                        {...register("telefono", {
-                                            required: "El telefono es un dato obligatorio",
-                                            pattern: {
-                                                value:
-                                                    /^\d{10}$/,
-                                                message:
-                                                    "El telefono debe tener el siguiente formato: 3865332311",
-                                            },
-                                        })} />
-                                    <Form.Text className="text-danger">
-                                        {errors.telefono?.message}
-                                    </Form.Text>
-                                </InputGroup>
-                            </Col>
-                            <Col xs={12} md={12} lg={12} className="mt-3" >
-                                <Form.Group>
-                                <Form.Label>Describa su consulta</Form.Label>
-                                    <Form.Control
-                                        as="textarea"
-                                        placeholder="Describa su inconveniente / consulta"
-                                        style={{ height: '100px' }}
-                                        {...register("consulta", {
-                                            required: "La consulta es un dato obligatorio",
-                                            minLength: {
-                                                value: 2,
-                                                message: "La cantidad minima de caracteres es de 2 digitos",
-                                            },
-                                            maxLength: {
-                                                value: 400,
-                                                message: "La cantidad maxima de caracteres es de 30 digitos",
-                                            },
-                                        })}
-                                    />
-                                    <Form.Text className="text-danger">
-                                        {errors.consulta?.message}
-                                    </Form.Text>
-                                </Form.Group>
-                            </Col>
-                        </Row>
-                            <Button variant="primary" type="submit" className="mt-3">
-                                Submit
-                            </Button>
-                    </Form>
-                </Col>
-            </Row>
-        </Container>
-    )
-}
