+++ src/components/views/Marcas.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
-import Imagen1 from "../assets/Marcas/eukanuba.png"
-import Imagen2 from "../assets/Marcas/Miproyecto.png"
-import Imagen3 from "../assets/Marcas/whiskas.png"
-import Imagen4 from "../assets/Marcas/proplan.png"
-import Imagen5 from "../assets/Marcas/vitalcan.png"
-import Imagen6 from "../assets/Marcas/sieger.png"
-import Imagen7 from "../assets/Marcas/agility.png"
-import Imagen8 from "../assets/Marcas/nupec.png"
-import Imagen9 from "../assets/Marcas/monello.png"
+import Imagen1 from "../../assets/Marcas/eukanuba.png"
+import Imagen2 from "../../assets/Marcas/Miproyecto.png"
+import Imagen3 from "../../assets/Marcas/whiskas.png"
+import Imagen4 from "../../assets/Marcas/proplan.png"
+import Imagen5 from "../../assets/Marcas/vitalcan.png"
+import Imagen6 from "../../assets/Marcas/sieger.png"
+import Imagen7 from "../../assets/Marcas/agility.png"
+import Imagen8 from "../../assets/Marcas/nupec.png"
+import Imagen9 from "../../assets/Marcas/monello.png"
 const Marcas = () => {
     return (
         <div className='d-flex justify-content-center align-items-center  slider'>
